自签名HTTP认证

+++ api/user/delete.go
@@ -11,6 +11,10 @@ import (
 
 // Delete delete an user by the user identifier.
 func Delete(c *gin.Context) {
+	if c.Request.Method != "DELETE" {
+		SendResponse(c, errno.ErrorUseDeleteMethod, nil)
+		return
+	}
 	userId, _ := strconv.Atoi(c.Param("id"))
 	if err := model.DeleteUser(uint64(userId)); err != nil {
 		SendResponse(c, errno.ErrDatabase, nil)
